Show more information on ScannerErrors
At the moment, Scanner Errors show relatively little information about what caused them.
All you see is the block that the Scanner Error occurred in. This change shows the
original error message (which was being hidden in a try-except), which points you directly
to the line that the Scanner Error occurred.

It's not ideal for newer users still, and could use a more guided approach based on the error
("i.e. it looks like your indentation is wrong"), but this is a first step.