Stop executing `objects from file` blocks each run
Since `objects from file` can't be mandatory, the `mark_as_answered`
function wasn't marking them as completed, and would be executed on each
step of the interview, like an `initial` block, and would recreate
objects that were already loaded and modified.

This change fixes that by directly adding the name of the question to
the internal "answered" dictionary.