Sets radio buttons in all cases
Currently docassemble can set the value of a radio button correctly. However,
in some PDF viewers (I tested in Ubuntu's Evince document viewer), the selected
radio button doesn't appear properly. This is because while docassemble
sets `/V` on the radio button group annotation, it doesn't set the `/AS`
(appearance stream) of the child annotation, which some viewers use to
actually which radio button is selected.

This PR makes the following changes to fix the issue:

* allows `export_values` to handle multiple values on the same field.
  For radio buttons, means that if a template already has a radio button
  selected, it won't override the given value in `data_strings`.
* Children of terminal fields don't have to have either FT or T, so
  look until a parent does and don't stop if the kid has FT, but not T
* keep track of when we have a child annotation of a `/Btn`, and make
  sure to set the appearance stream of that child to be the proper
  selected value

This PR is related to, but not dependent on, #619.
Some long standing discussion about this issue is in https://github.com/SuffolkLITLab/docassemble-ALWeaver/issues/75.