Remove cpp from workflow
(note: un-solicited contribution)

It seems that when adding [codeql](https://github.com/jhpyle/docassemble/commit/c4268e330fa504767ec87aa1aaae563aab48fca4),
GitHub detected `sample.cpp` in docassemble_demo, and thought that this repo
contained C++ code. This isn't correct, and since then, codeql complains loudly
that the C++ workflow fails (and sends me silly emails about it).

This removes that workflow. Feel free to clarify if the C++ workflow is
intentional and ignore this PR if that is the case.