Added correct line numbers in errors
There is no indication on existing error screens what line of the input YAML
file where the error is happening. This, combined with the fact that some error
messages have incorrect line numbers, is very confusing for Docassemble users
of all skills.

This PR adds the starting line number of the block for all errors. It
additionally corrects line numbers for `yaml.errors.MarkedYAMLError`s, which
are most errors emitted when parsing YAML.