Address codeql DOM text issue
There were a few issues in the CodeQL check, but this is the only one introduced
in this PR.

Instead of directly taking the `data-bs-theme` from the DOM, we can simply
check if the theme is dark, and use the string "dark". Otherwise, we use
"light", which prevents any direct DOM XSS issues.

https://cheatsheetseries.owasp.org/cheatsheets/DOM_based_XSS_Prevention_Cheat_Sheet.html

The full CodeQL run: https://github.com/jhpyle/docassemble/pull/718/checks?check_run_id=19234176736