Show detailed info about compiled code blocks
Currently, when a code block fails at runtime, the only piece of information
in the stack trace is `<code block>`, with no indication where the code block
was located.

This PR adds the `id_debug` function to Questions, and uses it to get info
about the file and line number of the code block. It puts this info in
compiled `validation code` and `code` blocks, so if they fail at runtime,
the error contains a relevant file and line number.