Make "Sign up" & "Sign in" similar w/ login links
When `login link style` is "normal" with a wide screen, the login links are
styled differently: one as a plain, unmarked link, and one as a button. This
change makes the links consistent on narrow screens and on wide screens: either
both links on narrow screens (in the dropdown) or both buttons on wider screens.

Additionally fixed an issue where there would be no login URL using login link
style and not allowing registration.