Allow `pyproject.toml`-only packages to be installed
Currently, if you have a project that only has a `pyproject.toml`,
you are able to install it from PyPI and from GitHub, but not as a
zip file. This is because docassemble only looks in `setup.py` to
find the package's name, which is needed to start the installation process
(server.py, `get_package_name_from_zip`).

With this PR, docassemble will look for the `project.name` entry in
a project's `pyproject.toml` if there is no `setup.py` present when
trying to install a package. This lets package developers store all their 
package config in a single file instead of multiple, and lets them
use modern packaging tools like `uv` locally.

This PR does __not__ yet add `pyproject.toml`-only support in the playground,
which would be a much bigger PR. If there's interest, I can start looking into
that.