Don't put extra set of quotes in `pyproject.toml`
When a `pyproject.toml` is generated from a playground project, an extra set of
single quotes inside of the double quotes is included around each dependency. This
is because single quotes are used in the `setup.py`, and they are included when
generating the dependency string for both.

This PR keeps the pyproject dependency list single quote free, and adds it to
the string that gets used in `setup.py`.