Find both `setup.py` and `pyproject.toml` if both are present
Currently, `get_package_name_from_zip` will find whichever of `setup.py` or `pyproject.toml`
are at the `min_level` (i.e. closest to the project root). However, due to how the comparison
is done, if they are at the same directory level, either one could be found and the other wouldn't be
(depending on which one is iterated over first).

Existing projects from
[docassemble 1.7](https://github.com/jhpyle/docassemble/blob/484736005270dd6107bf5535dc5dfd1d2e017fab/docassemble_webapp/docassemble/webapp/files.py#L648-L652)
have a limited `pyproject.toml` (only the `[build-system]`), and don't include the
name of the project. The name is necessary if only `pyproject.toml` is used. The only workaround is
to fill in all of the `[project]` section; if just the name is there, `pip` will require
all information from `setup.py` to be present.

Pip's exact error message (after adding `[project] name` and `version`):

```
The following seems to be defined outside of `pyproject.toml`:

`dependencies = [...]`

According to the spec (see the link below), however, setuptools CANNOT
consider this value unless `dependencies` is listed as `dynamic`.

https://packaging.python.org/en/latest/specifications/pyproject-toml/#declaring-project-metadata-the-project-table

To prevent this problem, you can list `dependencies` under `dynamic` or alternatively
remove the `[project]` table from your file and rely entirely on other means of
configuration.
```

This change (just a `<` to a `<=`) allows both `setup.py` and `pyproject.toml` to be found
if both are in the same directory level, and defers to the `setup.py` if it's present (as it's
more likely to be complete). This prevents any errors with installing projects that were
generated with docassemble 1.7. It also adds the exception message to the `api_package()` response
(which is needed to diagnose issues like this).