Change 'admin@admin.com' to 'admin@example.com'
Realized recently that `admin.com` is a real site. There
doesn't seem to be a mail server attached to it, but if the
owners decided to make one (or it's compromised), the
default admin account would be vulnerable.

`example.com` is an IANA reserved domain, meaning that it is not
available for registration (i.e. no one will ever own it).

https://www.iana.org/help/example-domains