Fix cloud convert output filename.
Got this warning email from Cloud convert when trying to setup docassemble to use it:

> For convert tasks, the filename parameter needs to always be the filename of the output file.
> It seems you have set filename to the name of the input file: For example, in task ID ..., you have set filename to myoutput.docx but you were converting to pdf.
> This results in errors and unexpected results

The fix is to just change the `filename` param in `cloudcovert_to_pdf` to `myoutput.pdf` instead of `myoutput.docx`.
The param is not used elsewhere, so nothing else needs to change.

It doesn't seem to break anything in docassemble, but it is technically incorrect.
Cloud Convert's email reads like it could break something on their side.