Added Basic Auth for Gotenberg servers
Gotenberg added basic auth recently, which is highly encourged when running it on a publically
visible URL. https://github.com/gotenberg/gotenberg/issues/684

This change adds `gotenberg: username` and `gotenberg: password` to the config to allow docassembl
to call gotenberg servers secured like this. It splits the config to still respect the original
`gotenberg url: ...` config or to handle the config being under an attribute, like such.

```yaml
gotenberg:
  url: ...
  username: ...
  password: ...
```

as well as adding the corresponding env vars `GOTENBERGUSERNAME` and `GOTENBERGPASSWORD`.

Tested that the original config works, as well as spinning up a new server with the gotenberg URL.