Fix custom objects loosing data on validation failure
Originally tracked in https://github.com/SuffolkLITLab/docassemble-ALToolbox/issues/131, finally
came back to it.

On each "continue", all of the filled values of objects are sent to the server and validated.
If any of the values are not valid, the user is shown the same page contents, with the values
they entered set to the `value` attribute of the given input, accomplished by setting the
defaults of those fields (server.py, lines 9169:9185).

However, if the field that needs to have the `value` attribute set is a custom object, it's
value won't be present, as it is deleted from the user dict (server.py, line 8213).

This PR fixes that behavior with two changes:

1. Giving each custom object a unique value in the user dict so more than one
  can exist at the same time (i.e. two custom fields on the same page).
2. Keeping those objects around in the user dictionary so they are present
  when the defaults are populated again.

  While I tried to remove them from the user dict after they are used, this
  doesn't seem to work, as the `__DANEWOBJECT....` values are still present in the user dict
  on subsequent pages. It's a little messy, but wouldn't break any aspect of existing
  interviews.

So now, if a user's entry in a custom field fails validation elsewhere on the page,
it won't force them to reenter the same value again.