Add Gotenberg params that work on Gotenberg 7.10 and newer
Since [version 7.10](https://github.com/gotenberg/gotenberg/releases/tag/v7.10.0),
Gotenberg has used the `pdfa` and `pdfua` form params instead of the
`nativePdfFormat` to generate PDF/A files. PDFs generated by Gotenberg
with the current params are normal, non archive PDFs.

Additionally, [Libreoffice no longer offers PDF/A-1a](https://gotenberg.dev/docs/troubleshooting#pdfa-1a)
as a target, as apparently that was always mislabeled and not really conformant. PDF/A-1b is the accepted
alternative (there's also PDF/A-2b, but figured sticking with PDF/A-1 was better for backwards
compatibility for folks).

This PR adds the new params to the gotenberg call, while leaving in the older param. On
the newer Gotenberg instances (tested on 8.23), the older param is ignored, and vice-versa
on the older gotenberg instances.