Fix default errors for `bates_number()`
If any of the `font_size`, `offset_horizontal`, and `offset_vertical`
params aren't given, then `bates_number` will error, because the default
values for those params are given as `None`, which do not parse into `floats`
on the `bates.py` side.

This patch changes the defaults in `util.py` to match the defaults in
`marisol.py`: 15 for the offsets and 10 for the font size.

Also changes how the values are retrieved from `kwargs`, to prevent people
from breaking the function by passing in `None`s.